+++ src/server/game/Entities/Pet/Pet.h
@@ -83,7 +83,7 @@ class Pet : public Guardian
         void setDeathState(DeathState s, bool despawn = false);                   // overwrite virtual Creature::setDeathState and Unit::setDeathState
         void Update(uint32 diff);                           // overwrite virtual Creature::Update and Unit::Update
 
-        uint8 GetPetAutoSpellSize() const { return m_autospells.size(); }
+        uint8 GetPetAutoSpellSize() const { return uint32(m_autospells.size()); }
         uint32 GetPetAutoSpellOnPos(uint8 pos) const
         {
             if (pos >= m_autospells.size())
